Update patch set 1
Patch Set 1:

The strange out of order and dropped message pieces issues aside, I see this mostly as a user experience issue.  With enableMultipartSMS=true, the message, although sent in 160 byte chunks, is reassembled at the receiving end so it appears as one long message to the user.  This is how (most? all? many?) modern stock ROMs seem to work so using the enableSplitSMS workaround in CM is a downgrade from the stock experience for the user, IMO.

Patch-set: 1