Update patch set 1
Patch Set 1:

Hmm, so multipart message I sent to both gmail and myself are received as split messages.  Maybe that's a sending limitation on CDMA?

Anecdotally the messages I send to myself are received slowly, but that might be a crummy network situation.

How is it working for other folks?

Patch-set: 1