Update patch set 2
Patch Set 2:

I'm not sure why the assumption is that cdma should use split sms.  This is not how 3gpp handles text. http://en.wikipedia.org/wiki/Concatenated_SMS

This is the config I use which is patched using the above dependency on d2vzw and it works perfectly. It converts to mms if its 4 messages long :

 <int name="smsToMmsTextThreshold">4</int>

    <!-- If true, The text message over 160 characters will be sent in multi part.
If false, The text message over 160 characters will be sent
via multi media message. -->
    <bool name="enableMultipartSMS">true</bool>

    <!-- If true, the text message will be split every 160 characters.
If false, the text will never be split before being sent. -->
    <bool name="enableSplitSMS">false</bool>

Patch-set: 2