Update patch set 2
Patch Set 2:

Correct.  If I send a message with the above settings and http://review.cyanogenmod.com/#/c/22356/ , I can send a multipart text message and it is not split (at least visibly to the user) at sending or receiving.  I have verified that this works with mms.apk and handcent on a d2vzw.  None of my messages went to a Sprint user and I don't know how they typically send and receive messages.

Patch-set: 2