Update patch set 1
Patch Set 1:

Something in A2220 is causing the call delay for us. (When you answer a phone call it takes 3-5 seconds before you can hear or they can hear you.)

Commenting this out seems to fix it, but I'm not sure that's the proper thing to do here.

Just a heads up.

Patch-set: 1