Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

Aligns with pre and post conditions with the ril. This is for the older version of the RIL blobs.
Having the wrong value will result unintended problem
revert when we get the final v7 ril blobs which is designed for jellybean.
 
Shipping it now.

Patch-set: 1