Update patch set 1
Patch Set 1:

We're still getting occasional screen flicker with 3-layer overlays.  I can reliably reproduce it with Opera + Screen Filter, but there's other circumstances in which it happens.  bbelos has run into it with Google Music and the volume box.

Given that we're unlikely to debug it before the next M-release (especially if it's kernel related), should we consider a 2-layer limit so it's not user visible?

Threw in d2-common as we suspect it may be kernel related and don't want to impose it on other (theoretical) devices, but I can move to qcom-common if that's more appropriate.

Patch-set: 1