Update patch set 1
Patch Set 1:

OK I reproduced it with Opera and Screen Filter. I had not seen this before, so there was nothing to debug. I'll take care of it. Switching to 2-layer is not the right thing to do and I'd prefer not to release at all than to do that.

Patch-set: 1