Update patch set 1
Patch Set 1:

The flicker itself comes and goes.  But overlay is only used when the screen contents are static (i.e., when you stop scrolling or something).  It does happen on the intro screen but it happens on a variety of pages too.

I've also seen it happen in the launcher and Clock, but it's harder to reproduce.

Apologies, I thought we had discussed this issue before, but I guess that was just about the previous artifacting.

Patch-set: 1