Update patch set 1
Patch Set 1:

USB Storage mounts and is accessible (still acts retarded and mounts/unmounts as an sdcard prompting to replace after unmounting and removing)

Mainly....gets rid of the com.android.media crash that /storage/usbdisk0 wasn't found.

Seems also, there is a permission issue to these dirs in init for this elsewhere as well when no device is connected..keeping this patch small however.

Patch-set: 1