Update patch set 1
Patch Set 1:

Few notes:

Camcorder is essentially "Handset DMIC Stereo" (which matches "Stereo Mic" in TW) but with the L & R channels flipped so they're mapped properly when using the camcorder in the typical (90°?) landscape orientation.  Of course, if it's rotated the other way (270°?), the channels will be reversed.

The speakerphone echo I had were "blips" of the caller voice going through at the end of the caller saying something.  Changing the ACDBID seems to fix that--I flipped back and forth a few times and it was consistent.

Uncertain if the "Voice Recognition" ACDBID really matters, but it was a discrepancy I found against stock.

Patch-set: 1