Update patch set 1
Patch Set 1:

Yeah, in CM 10.0 the camcorder records in mono using the Sub Mic (the microphone located on the top of the device when held in portrait orientation).

With this the camcorder records in stereo, with the sub/top mic as the left channel, and the bottom/main mic as the right channel.

Patch-set: 1