Update patch set 1
Patch Set 1: Do not submit

Ehhh

Did you actually test this?

Android's initlogo will write garbage because the display expects it in a different format. The splash is currently done by opening the framebuffer (in kernelspace) during the device probe and drawing the logo. This is totally unrelated- it's disabled because the kernel already does it.

There's a race condition in the kernel which causes a lockup during boot occasionally because of the new non-blocking update code. I'll fix it this week.

Patch-set: 1