Update patch set 3
Patch Set 3:

Yes, fully sync'd.. using a d2vzw and Samsung multimedia desktop dock for the Samsung Infuse 4G.

If you try to adjust the volume, the slider shows up but doesn't move with inputs and the slider color is dimmer than usual.

I tried to grab a screenshot but the screenshot action clears the volume popup.

Patch-set: 3