Update patch set 1
Patch Set 1:

There's some weirdness with the Tethering checkbox in settings - after you press it, it does not produce a check-mark.  Regardless, with this patch, the interface is kicked off, and you are able to tether via USB.  However because the tick-mark is not there, the only way you can stop tethering is to disconnect the USB connection.

Patch-set: 1