Update patch set 1
Patch Set 1:

Supports the new API used by versions of libloc_api_v02.so that have shipped in recent d2spr updates, which are otherwise incompatible with the existing adapter code.

Requires the new blob, located at the usual place.  AFAIK this version of the blob uses the most recent CAF sources (of blobs actually shipping), and includes the privacy lock functionality.  The blob we're presently using, in addition to not supporting the privacy lock, seems oddly out of date too.

Are the struct changes in include/hardware/gps.h still necessary?  I'm not sure how to properly test those.

With regard to the privacy lock, I'm not sure there's reason to disable it as long as it still works on d2att/d2tmo.  This gives us the benefit of not having to use device-specific blobs.

Patch-set: 1