Update patch set 1
Patch Set 1:

Right now we have two wcn devices inheriting off d2-common. (Believe it or not it works, with a lot of ifdefs) If I submit this I'm going to move the product_packages declaration to the device tree so it doesn't build for d2. noobnl suggested we keep the scripts and code in d2-common though until/if we move to another common repo for these devices.

Patch-set: 1