Update patch set 1
Patch Set 1:

We'll use the stock-ish logo in the interim while we design a new one that follows the rules (using the freexperia logo as a guide). In the meantime, can I get this merged? the d2 logo is the wrong size for our screen and messes up the framebuffer

Patch-set: 1