Update patch set 1
Patch Set 1:

Specifically I'm not sure.  Really, I just noticed that they were different than the entries in our stock build.prop.  I don't know if they were intentionally modified or just transposed incorrectly when somebody first set this up.  Or maybe that's what was there when the first d2 was released and was never bothered to be modified later (as in the radio.snapshot.enabled/disabled).

The other one I believe was a 'fix' for when we were having problems with NV/RUIM settings sticking.  I believe that issue was fixed elsewhere but the prop was never reverted back to its previous state.

Its NBD if it doesn't get merged really, if its actually causing some reproducible problems with these turned on and off.  Turn your phone on and off 10x's with and 10x's without and see exactly if there is much difference.  I would think if it were an issue, the users running my unofficial builds (over 1100 right now) would be complaining.

Patch-set: 1