Update patch set 1
Patch Set 1:

I have the same amount of users running with similar tweaks for the last week or so over on d2att, and nobody has stated anything. I have noticed these issues myself, and noticed that they dissipated once I changed them. The phone boots, but if you lock the screen once you turn on the phone it gets stuck in a locked state

Patch-set: 1