Update patch set 1
Patch Set 1:

I just rebooted, waited for the boot animation to finish then locked, woke the phone up, unlocked a few times before it was finally finished booting.

Have you tried only modifying one of the lines at a time to see which might be causing it?  If I had to guess it would be the no_wait_for_card line... but I don't seem to have the issue you describe with it as is.

Patch-set: 1