Update patch set 3
Patch Set 3:

The actual device name from the stock is d2lterefreshspr.

Also with the modem L710TVPU0ANF4 for the L710T the symlinks need to not substitue modem_fw to modem_f2 like all of the other d2 variants in order for the baseband to work.  

This solution is working for me : http://review.teamoctos.com:8080/#/c/511/1/Android.mk

along with device d2lterefreshspr 
http://review.teamoctos.com:8080/#/c/512/

Patch-set: 3