Update patch set 1
Patch Set 1:

The reboots also occurs on d2tmo. The only time it happens for me at least is when it tries to connect to cellular data but can't establish a connection because of weak signal(Happened to me twice when I was on the bus without WIFI available). So what happens is that it eventually gives up trying to connect and reboots. But the reboot is much worst. It reboots and begins to optimize the apps again, then once that is done, it reboots again and begins the same process. The only fix to this is to pull out the battery and put it back it and let it optimize the apps again, but this time it will boot correctly.

Patch-set: 1