Update patch set 1
Abandoned

keep this on topic please. Your reboots have nothing to do with the proposal to add swap ram. And I've already seen enough of these logs. I'm tired of the constant requests from matrix zone if I'm fixing the problem as well. If I had a solution I would have pushed it to gerrit already. Do you seriously think I would either not fix the problem or fix it and keep the solution to myself?

They are nightlies, they are expected to be a bit rough. If you want stability stick to M releases.

Patch-set: 1
Status: abandoned