Update patch set 2
Patch Set 2:

I flashed a d2vzw nightly with this Change and the a2dp audio is back, it stutters a lot, but it's actually playing something, which is a great start! Testing seems to revel that apps running in the background that have permissions to view bt profiles might be causing high CPU, which in turn might be causing the stuttering. Also, I tested a call over bluetooth and that works so I'm not sure if it breaks that functionality as Arne described a few comments above. I will continue to test this and see if I can find anything else out.

Patch-set: 2
Reviewer: Mike Yabby <15126@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0