Update patch set 17
Patch Set 17:

Couldnt test hangouts calls, my wifi was being a bitch :/
I tested music and audio recording on speaker/phone mic, headphones without mic, and headphones with mic. On headphones without mic, on audio playback, it only comes out of one ear. I'll get a log, and go back to the old hal to see if the issue is present there. Also, Dave and Dan, BT turns on and works for me with this, but I can't test BT audio

Patch-set: 17