Update patch set 17
Patch Set 17:

both ears work on the old hal, but at first it seemed like only one ear worked, but when I hit play again it was both ears so I could've been imagining it or it could've been the song. grabbed a log, uploading both, and emailing them to you arne.

Patch-set: 17